Define public API for `url_operations`
This implements the plan declared at
https://github.com/datalad/datalad-next/issues/613#issuecomment-1923489248

A dedicated list of components is exposed at the package's top-level,
and only that. All other functionality and imports have been move to
package-internal files.

All other, now undesired, top-level imports have been annotated and
scheduled for removal with the v2.0 release.
